pyproject.toml
[project]
name = "gxy-sketches"
version = "0.1.0"
description = "Distill open bioinformatics workflows (nf-core, Galaxy IWC) into skill-like markdown sketches for LLM agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.7",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "httpx>=0.27",
]

[project.scripts]
gxy-sketches = "gxy_sketches.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gxy_sketches"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
